A `NSSPCredentials` Class Representing a Credentials object

A `Credentials` object has a username, a password and a key.

Details

A `Credentials` object can get ESSENCE API data via an API URL.

Methods

Public methods

Credentials$new()

Credentials$get_api_response()

Credentials$get_api_data()

Credentials$get_api_tsgraph()

Credentials$clone()

Method new()

Initializes a new Credentials object.

Usage

Credentials$new(username, password)

Arguments

username

a character of ESSENCE username

password

a character of ESSENCE password

Returns

A new `Credentials` object

Method get_api_response()

Get ESSENCE API response

Usage

Credentials$get_api_response(url)

Arguments

url

a character of ESSENCE API URL

Returns

An object of class response

Examples

\dontrun{
myProfile <- Credentials$new(askme("Enter my username: "), askme())
url <- "<ESSENCE_url>"
api_response <- myProfile$get_api_response(url)
}

Method get_api_data()

Get ESSENCE API data

Usage

Credentials$get_api_data(url, fromCSV = FALSE)

Arguments

url

a character of ESSENCE API URL

fromCSV

a logical, defines whether data are returned in .csv format or .json format

Returns

A dataframe (`fromCSV = TRUE`) or a list containing a dataframe and its metadata (`fromCSV = FALSE`)

Examples

\dontrun{
myProfile <- Credentials$new(askme("Enter my username: "), askme())
json_url <- "<json type ESSENCE_url>"
api_data_json <- myProfile$get_api_data(json_url)

csv_url <- "<csv type ESSENCE_url>"
api_data_csv <- myProfile$get_api_data(csv_url, fromCSV = TRUE)
}

Method get_api_tsgraph()

Get ESSENCE API data

Usage

Credentials$get_api_tsgraph(url)

Arguments

url

a character of ESSENCE API URL

Returns

A list containing an api_response object and a path to a time series graph in .png format

Examples

\dontrun{
myProfile <- Credentials$new(askme("Enter my username: "), askme())
url <- "<ESSENCE_url>"
api_data_tsgraph <- myProfile$get_api_tsgraph(url)
names(api_data_tsgraph)
img <- png::readPNG(api_data_tsgraph$tsgraph)
grid::grid.raster(img)
}

Method clone()

The objects of this class are cloneable with this method.

Usage

Credentials$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Credentials$get_api_response`
## ------------------------------------------------

if (FALSE) {
myProfile <- Credentials$new(askme("Enter my username: "), askme())
url <- "<ESSENCE_url>"
api_response <- myProfile$get_api_response(url)
}

## ------------------------------------------------
## Method `Credentials$get_api_data`
## ------------------------------------------------

if (FALSE) {
myProfile <- Credentials$new(askme("Enter my username: "), askme())
json_url <- "<json type ESSENCE_url>"
api_data_json <- myProfile$get_api_data(json_url)

csv_url <- "<csv type ESSENCE_url>"
api_data_csv <- myProfile$get_api_data(csv_url, fromCSV = TRUE)
}

## ------------------------------------------------
## Method `Credentials$get_api_tsgraph`
## ------------------------------------------------

if (FALSE) {
myProfile <- Credentials$new(askme("Enter my username: "), askme())
url <- "<ESSENCE_url>"
api_data_tsgraph <- myProfile$get_api_tsgraph(url)
names(api_data_tsgraph)
img <- png::readPNG(api_data_tsgraph$tsgraph)
grid::grid.raster(img)
}