Exponentially Weighted Moving Average (EWMA)

The EWMA compares a weighted average of the most recent visit counts to a baseline expectation. For the weighted average to be tested, an exponential weighting gives the most influence to the most recent observations. This algorithm is appropriate for daily counts that do not have the characteristic features modeled in the regression algorithm. It is more applicable for Emergency Department data from certain hospital groups and for time series with small counts (daily average below 10) because of the limited case definition or chosen geographic region.

alert_ewma(df, t = date, y = count, B = 28, g = 2, w = 0.4)

Arguments

df
A data frame, data frame extension (e.g. a tibble), or a lazy data frame.

t
Name of the column of type Date containing the dates

y
Name of the column of type Numeric containing counts or percentages

B
Baseline parameter. The baseline length is the number of days from which to compute the EWMA and compare to in order to form the test statistic (default is 28)

g
Guardband paramter. The guardband length is the number of days separating the baseline from the current date in consideration for alerting (default is 2)

w
Smoothing parameter. The smoothing parameter controls the amount of smoothing in the exponentially weighted moving average (default is 0.4)

Value

A data frame

Examples

# Example 1
df <- data.frame(date = seq.Date(as.Date("2020-01-01"), as.Date("2020-12-31"), by = 1),
                 count = floor(runif(366, min=0, max=101)))
df_ewma <- alert_ewma(df)

head(df)

#>         date count
#> 1 2020-01-01     8
#> 2 2020-01-02    84
#> 3 2020-01-03    60
#> 4 2020-01-04    15
#> 5 2020-01-05     0
#> 6 2020-01-06    47

head(df_ewma)

#>          date count          t  y       mu        s        z test_statistic
#> 1: 2020-01-31     3 2020-01-31  3 44.57143 30.81057 38.08712   -0.383510999
#> 2: 2020-02-01    22 2020-02-01 22 46.78571 30.30873 31.65227   -0.909879234
#> 3: 2020-02-02    30 2020-02-02 30 46.25000 29.75408 30.99136   -0.934507694
#> 4: 2020-02-03    64 2020-02-03 64 44.21429 30.71296 44.19482   -0.001155092
#> 5: 2020-02-04    48 2020-02-04 48 44.46429 30.49406 45.71689    0.074853581
#> 6: 2020-02-05    43 2020-02-05 43 45.53571 29.38062 44.63013   -0.056166784
#>      p.value alert
#> 1: 0.3521721  blue
#> 2: 0.1854688  blue
#> 3: 0.1791626  blue
#> 4: 0.4995434  blue
#> 5: 0.4704415  blue
#> 6: 0.4778113  blue

# Example 2
df <- data.frame(Date = seq.Date(as.Date("2020-01-01"), as.Date("2020-12-31"), by = 1),
                 percent = runif(366))
df_ewma <- alert_ewma(df, t = Date, y = percent)

head(df)

#>         Date    percent
#> 1 2020-01-01 0.93648743
#> 2 2020-01-02 0.97068136
#> 3 2020-01-03 0.71027405
#> 4 2020-01-04 0.88509093
#> 5 2020-01-05 0.97687532
#> 6 2020-01-06 0.03489964

head(df_ewma)

#>          Date    percent          t          y        mu   s         z
#> 1: 2020-01-31 0.89274570 2020-01-31 0.89274570 0.6138711 0.5 0.7007241
#> 2: 2020-02-01 0.84387029 2020-02-01 0.84387029 0.6115889 0.5 0.7579826
#> 3: 2020-02-02 0.72976121 2020-02-02 0.72976121 0.5905639 0.5 0.7466940
#> 4: 2020-02-03 0.03841498 2020-02-03 0.03841498 0.5970808 0.5 0.4633824
#> 5: 2020-02-04 0.33382332 2020-02-04 0.33382332 0.5956086 0.5 0.4115588
#> 6: 2020-02-05 0.74927265 2020-02-05 0.74927265 0.5867831 0.5 0.5466443
#>    test_statistic   p.value alert
#> 1:      0.3165399 0.3770136  blue
#> 2:      0.5335386 0.2990130  blue
#> 3:      0.5690236 0.2870218  blue
#> 4:     -0.4872702 0.3149998  blue
#> 5:     -0.6707786 0.2540298  blue
#> 6:     -0.1462878 0.4423905  blue

if (FALSE) {
# Example 3: Data from NSSP-ESSENCE
library(ggplot2)

myProfile <- Credentials$new(askme("Enter your username:"), askme())

url <- "https://essence2.syndromicsurveillance.org/nssp_essence/api/timeSeries?endDate=20Nov20
&percentParam=ccddCategory&datasource=va_hosp&startDate=22Aug20&medicalGroupingSystem=essencesyndromes
&userId=2362&aqtTarget=TimeSeries&ccddCategory=cli%20cc%20with%20cli%20dd%20and%20coronavirus%20dd%20v2
&geographySystem=hospitalstate&detector=probregv2&timeResolution=daily&hasBeenE=1&stratVal=
&multiStratVal=geography&graphOnly=true&numSeries=0&graphOptions=multipleSmall&seriesPerYear=false
&nonZeroComposite=false&removeZeroSeries=true&sigDigits=true&startMonth=January&stratVal=&multiStratVal=geography
&graphOnly=true&numSeries=0&graphOptions=multipleSmall&seriesPerYear=false&startMonth=January&nonZeroComposite=false"

url <- url %>% gsub("\n", "", .)

api_data <- myProfile$get_api_data(url)

df <- api_data$timeSeriesData

df_ewma <- alert_ewma(df, t = date, y = count)

# Visualize alert for South Dakota
df_ewma_state <- df_ewma %>%
  filter(hospitalstate_display == "South Dakota")

df_ewma_state %>%
  ggplot(aes(x = t, y = count)) +
  geom_line(color = "blue") +
  geom_point(data = subset(df_ewma_state, alert == "red"), color = "red") +
  geom_point(data = subset(df_ewma_state, alert == "yellow"), color = "yellow") +
  theme_bw() +
  labs(x = "Date",
       y = "Percent")
}