Multiple Adaptive Regression

The multiple adaptive regression algorithm fits a linear model to a baseline of counts or percentages of length B, and forecasts a predicted value g + 1 days later (guard-band). This value is compared to the current observed value and divided by the standard error of prediction in the test-statistic. The model includes terms to account for linear trends and day-of-week effects. Note that this implementation does NOT include holiday terms as in the Regression 1.2 algorithm in ESSENCE. An alert is signaled if the statistical test (student's t-test) applied to the test statistic yields a p-value less than 0.01. If the p-value is greater than or equal to 0.01 and strictly less than 0.05, a warning is signaled.

alert_mar(df, t = date, y = count, B = 28, g = 2)

Arguments

df
A data frame, data frame extension (e.g. a tibble), or a lazy data frame.

t
Name of the column of type Date containing the dates

y
Name of the column of type Numeric containing counts or percentages

B
Baseline parameter. The baseline length is the number of days to which each liner model is fit (default is 28)

g
Guardband paramter. The guardband length is the number of days separating the baseline from the current date in consideration for alerting (default is 2)

Value

A data frame

Examples

# Example 1
df <- data.frame(date = seq.Date(as.Date("2020-01-01"), as.Date("2020-12-31"), by = 1),
                 count = floor(runif(366, min=0, max=101)))
df_mar <- alert_mar(df)

head(df)

#>         date count
#> 1 2020-01-01     4
#> 2 2020-01-02    80
#> 3 2020-01-03    20
#> 4 2020-01-04     6
#> 5 2020-01-05    87
#> 6 2020-01-06     7

head(df_mar)

#> # A tibble: 6 x 17
#>   date       count t              y   row initial_vals   Wed   Thu   Fri   Sat
#>   <date>     <dbl> <date>     <dbl> <int> <chr>        <dbl> <dbl> <dbl> <dbl>
#> 1 2020-01-31    36 2020-01-31    36    31 no               0     0     1     0
#> 2 2020-02-01    62 2020-02-01    62    32 no               0     0     0     1
#> 3 2020-02-02    46 2020-02-02    46    33 no               0     0     0     0
#> 4 2020-02-03    16 2020-02-03    16    34 no               0     0     0     0
#> 5 2020-02-04    51 2020-02-04    51    35 no               0     0     0     0
#> 6 2020-02-05    14 2020-02-05    14    36 no               1     0     0     0
#> # ... with 7 more variables: Sun <dbl>, Mon <dbl>, Tue <dbl>, fitted <dbl>,
#> #   statistic <dbl>, p.value <dbl>, alert <chr>

# Example 2
df <- data.frame(Date = seq.Date(as.Date("2020-01-01"), as.Date("2020-12-31"), by = 1),
                 percent = runif(366))
df_mar <- alert_mar(df, t = Date, y = percent)

head(df)

#>         Date   percent
#> 1 2020-01-01 0.2433926
#> 2 2020-01-02 0.2829591
#> 3 2020-01-03 0.2166260
#> 4 2020-01-04 0.1200414
#> 5 2020-01-05 0.1702637
#> 6 2020-01-06 0.5215289

head(df_mar)

#> # A tibble: 6 x 17
#>   Date       percent t                y   row initial_vals   Wed   Thu   Fri
#>   <date>       <dbl> <date>       <dbl> <int> <chr>        <dbl> <dbl> <dbl>
#> 1 2020-01-31 0.00408 2020-01-31 0.00408    31 no               0     0     1
#> 2 2020-02-01 0.657   2020-02-01 0.657      32 no               0     0     0
#> 3 2020-02-02 0.151   2020-02-02 0.151      33 no               0     0     0
#> 4 2020-02-03 0.199   2020-02-03 0.199      34 no               0     0     0
#> 5 2020-02-04 0.240   2020-02-04 0.240      35 no               0     0     0
#> 6 2020-02-05 0.662   2020-02-05 0.662      36 no               1     0     0
#> # ... with 8 more variables: Sat <dbl>, Sun <dbl>, Mon <dbl>, Tue <dbl>,
#> #   fitted <dbl>, statistic <dbl>, p.value <dbl>, alert <chr>

if (FALSE) {
# Example 3: Data from NSSP-ESSENCE
library(ggplot2)

myProfile <- Credentials$new(askme("Enter your username:"), askme())

url <- "https://essence2.syndromicsurveillance.org/nssp_essence/api/timeSeries?endDate=20Nov20
&percentParam=ccddCategory&datasource=va_hosp&startDate=22Aug20&medicalGroupingSystem=essencesyndromes
&userId=2362&aqtTarget=TimeSeries&ccddCategory=cli%20cc%20with%20cli%20dd%20and%20coronavirus%20dd%20v2
&geographySystem=hospitalstate&detector=probregv2&timeResolution=daily&hasBeenE=1&stratVal=
&multiStratVal=geography&graphOnly=true&numSeries=0&graphOptions=multipleSmall&seriesPerYear=false
&nonZeroComposite=false&removeZeroSeries=true&sigDigits=true&startMonth=January&stratVal=&multiStratVal=geography
&graphOnly=true&numSeries=0&graphOptions=multipleSmall&seriesPerYear=false&startMonth=January&nonZeroComposite=false"

url <- url %>% gsub("\n", "", .)

api_data <- myProfile$get_api_data(url)

df <- api_data$timeSeriesData

df_mar <- alert_mar(df, t = date, y = count)

# Visualize alert for South Dakota
df_mar_state <- df_mar %>%
  filter(hospitalstate_display == "South Dakota")

df_mar_state %>%
  ggplot(aes(x = t, y = count)) +
  geom_line(color = "blue") +
  geom_point(data = subset(df_mar_state, alert == "red"), color = "red") +
  geom_point(data = subset(df_mar_state, alert == "yellow"), color = "yellow") +
  theme_bw() +
  labs(x = "Date",
       y = "Percent")
}