US Native American Lands Shapefile

Useful `sf`` object shapefile for US Native American Lands.

tribal_sf

Format

An object of class sf (inherits from data.frame) with 770 rows and 22 columns.

Details

Rows: 770
Columns: 22
$ OBJECTID   <fct> 1522, 1523, 1524, 1525, 1526, 1527, 1528, 1529, 1530, 1531, 1532, 1533, 1534, 1535, 1536, 1537, 1538, 1539, 1540, 1541, 1542, 1543...
$ AIANNHCE   <fct> 2320, 9260, 2055, 1195, 1970, 4180, 1625, 0240, 0325, 3320, 4200, 1370, 1960, 1365, 1575, 3750, 0050, 0165, 0750, 0780, 1250, 1170...
$ AIANNHNS   <fct> 02419022, 01472036, 00647264, 00260277, 00275865, 02419492, 01934338, 00219211, 02419316, 00979449, 00023763, 02418951, 02418998, ...
$ GEOID      <fct> 2320R, 9260R, 2055R, 1195R, 1970R, 4180R, 1625R, 0240R, 0325R, 3320R, 4200R, 1370R, 1960R, 1365R, 1575R, 3750R, 0050R, 0165R, 0750...
$ NAME       <fct> "Mohegan", "Pamunkey", "Lower Sioux", "Fort Independence", "Lone Pine", "Timbi-Sha Shoshone", "Isleta", "Big Lagoon", "Blue Lake",...
$ NAMELSAD   <fct> "Mohegan Reservation", "Pamunkey (state) Reservation", "Lower Sioux Indian Community", "Fort Independence Reservation", "Lone Pine...
$ LSAD       <fct> 86, 86, 95, 86, 86, 86, 84, 85, 85, 86, 86, 86, 86, 82, 86, 85, 86, 86, 85, 97, 96, 86, 85, 85, 85, 86, 85, 85, 86, 86, 86, 86, 86...
$ CLASSFP    <fct> D8, D4, D2, D2, D2, D8, D2, D2, D8, D2, D8, D8, D8, D8, D2, D8, D8, D8, D2, D2, D8, D8, D2, D2, D2, D4, D2, D2, D2, D2, D8, D2, D2...
$ COMPTYP    <fct> R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R, R...
$ AIANNHR    <fct> F, S, F, F, F, F, F, F, F, F, F, F, F, F, F, F, F, F, F, F, F, F, F, F, F, S, F, F, F, F, F, F, F, F, F, F, F, S, F, F, F, F, F, F...
$ MTFCC      <fct> G2101, G2101, G2101, G2101, G2101, G2101, G2101, G2101, G2101, G2101, G2101, G2101, G2101, G2101, G2101, G2101, G2101, G2101, G210...
$ FUNCSTAT   <fct> A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A...
$ ALAND      <dbl> 1767260, 4421855, 6962148, 2261362, 952613, 31762611, 855026227, 21072, 136714, 49055456, 11533940550, 74273, 702364, 41539443, 97...
$ AWATER     <dbl> 0, 1958558, 124999, 0, 0, 0, 2052388, 0, 0, 5306039, 883039, 0, 7307, 0, 17867564, 0, 83761, 0, 0, 0, 35225904, 14695, 0, 0, 41930...
$ INTPTLAT   <fct> +41.4900915, +37.5718327, +44.5308483, +36.8411177, +36.5968832, +37.4456367, +34.8913841, +41.1642587, +40.8833451, +44.9787065, ...
$ INTPTLON   <fct> -072.0940543, -077.0052176, -094.9905189, -118.2289361, -118.0560422, -117.5247411, -106.6662027, -124.1207880, -123.9980188, -074...
$ Published_ <fct> 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018, 2018...
$ SHAPE_Leng <dbl> 0.13026724, 0.11004068, 0.21177248, 0.08093609, 0.05008642, 0.56260103, 1.93765470, 0.00720691, 0.01951197, 0.43234518, 6.96606854...
$ IHS_Area   <chr> "NASHVILLE", "NASHVILLE", "BEMIDJI", "CALIFORNIA", "CALIFORNIA", "CALIFORNIA,PHOENIX", "ALBUQUERQUE", "CALIFORNIA", "CALIFORNIA", ...
$ Shape_Le_1 <dbl> 0.130267238, 0.110040677, 0.211772478, 0.080936093, 0.050086421, 0.562601027, 1.937654698, 0.007206907, 0.019511970, 0.432345176, ...
$ Shape_Area <dbl> 1.905319e-04, 6.507490e-04, 8.023517e-04, 2.284535e-04, 9.593736e-05, 3.221397e-03, 8.451501e-02, 2.260852e-06, 1.460647e-05, 6.20...
$ geometry   <MULTIPOLYGON [m]> MULTIPOLYGON (((2269262 266..., MULTIPOLYGON (((1997398 -54..., MULTIPOLYGON (((394546 -387..., MULTIPOLYGON (((-1611...

Examples

library(Rnssp)
data("tribal_sf")

plot(tribal_sf$geometry)